Remove unused code and dependencies to Maven
These are no longer needed because all of the old MavenPlugin
functionality was removed. This removes all dependencies not needed
for the core of Gradle. And upgrades to the latest Maven version.

The Maven support in 'gradle init' (maven2gradle) still requires more
things from Maven which will be handled in a follow up.